Unify temporary netrc handling
